Backend/myapp/requirements.txt
Django==4.1.4
ipdb
pip
# django-autoreload==1.5.0
mysqlclient==2.1.0
djangorestframework==3.13.1
django-cors-headers
django-cors-middleware
django-storages==1.13.2
minio
boto3==1.26.68
python-dotenv
Pillow